requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchainhub
langchain-core
chromadb
httpx
pydantic
tqdm
sentence-transformers
openrouter
langchain-huggingface
openai
pypdf

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
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
